+++ demisto_sdk/commands/common/hook_validations/integration.py
@@ -103,6 +103,7 @@ class IntegrationValidator(ContentEntityValidator):
             self.is_valid_fetch(),
             self.is_there_a_runnable(),
             self.is_valid_display_name(),
+            self.is_valid_display_name_for_siem(),
             self.is_valid_pwsh(),
             self.is_valid_image(),
             self.is_valid_max_fetch_and_first_fetch(),
@@ -1031,6 +1032,18 @@ class IntegrationValidator(ContentEntityValidator):
                     return False
 
             return True
+
+    def is_valid_display_name_for_siem(self) -> bool:
+        is_siem = self.current_file.get('script', {}).get('isFetchEvents')
+
+        if is_siem:
+            display_name = self.current_file.get('display', '')
+            if not display_name.endswith('Event Collector'):
+                error_message, error_code = Errors.invalid_siem_integration_name(display_name)
+                if self.handle_error(error_message, error_code, file_path=self.file_path):
+                    return False
+
+        return True
 
     def is_valid_hidden_params(self) -> bool:
         """